requirements.txt
fastapi
uvicorn
streamlit

langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq

groq

faiss-cpu

sentence-transformers

pypdf
python-dotenv
requests
yfinance